Add switch for `reference` to point at IGS20
The new data is processed with IGS20 reference, so some urls are breaking for now.
These urls will change in the future.

Notes from UNR:
Links on the station pages will be updated to point to the IGS20 products. IGS14 products will be frozen but still available up to August 24, 2024. There may be a few issues or glitches during this transition over the next week or so, but we will make every attempt to minimize disruption.

Once the change goes into effect, the directories which contain the files will be reorganized. A map of the new directory structure is given HERE. Importantly, the old IGS14 solutions will be moved to be under https://geodesy.unr.edu/gps_timeseries/IGS14.